Move deploy job material into own module
