+++ src/main.rs
@@ -1,3 +1,4 @@
+use crate::deploy_job::DeployJob;
 use kube::api::{DeleteParams, Informer, ListParams, PatchParams};
 use kube::{
     api::{Api, Object, PostParams, WatchEvent},
@@ -9,6 +10,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use std::collections::HashMap;
 
+mod deploy_job;
 #[cfg(test)]
 mod tests;
 
@@ -206,92 +208,20 @@ fn start_gordo_deploy_job(
     namespace: &str,
     env_config: &GordoEnvironmentConfig,
 ) -> () {
-    let gordo_config = serde_json::to_string(&gordo.spec.config).unwrap();
 
-    // Create the job name.
-    let job_name_suffix = format!(
-        "{}-{}",
-        &gordo.metadata.name,
-        &gordo.metadata.generation.map(|v| v as u32).unwrap_or(0)
-    );
-    let job_name = deploy_job_name("gordo-dpl-", &job_name_suffix);
-
-    // Define the owner reference info
-    let owner_ref = json!([
-        {"blockOwnerDeletion": true, "uid": gordo.metadata.uid, "apiVersion": "v1", "kind": "Gordo", "name": &gordo.metadata.name, "controller": true }
-    ]);
-    let owner_ref_as_string = serde_json::to_string(&owner_ref).unwrap();
-
-    // TODO: Remove this after a few weeks/months when people have migrated >= 0.33 of gordo-deploy
-    let gordo_deploy_key_val = if minor_version(&gordo.spec.deploy_version) >= Some(33) {
-        json!({"name": "GORDO_NAME", "value": &gordo.metadata.name})
-    } else {
-        json!({"name": "MACHINE_CONFIG", "value": gordo_config})
-    };
-
-    // Build up the gordo-deploy environment variables
-    let mut env = vec![
-        gordo_deploy_key_val,
-        json!({"name": "ARGO_SUBMIT", "value":  "true"}),
-        json!({"name": "WORKFLOW_GENERATOR_PROJECT_NAME", "value": &gordo.metadata.name}),
-        json!({"name": "WORKFLOW_GENERATOR_OWNER_REFERENCES", "value": owner_ref_as_string}),
-    ];
-    // push in any that were supplied by the Gordo.spec.gordo_environment mapping
-    gordo.spec.deploy_environment.as_ref().map(|environment| {
-        environment.iter().for_each(|(key, value)| {
-            env.push(json!({"name": key, "value": value}));
-        })
-    });
-
-    // Job spec for launching this gordo config into a workflow
-    let spec = json!({
-        "apiVersion": "batch/v1",
-        "kind": "Job",
-        "metadata": {
-            "name": &job_name,
-            "ownerReferences": owner_ref,
-            "labels": {
-                "gordoProjectName": &gordo.metadata.name
-            }
-        },
-        "spec": {
-            "template": {
-                "metadata": {
-                    "name": &job_name
-                },
-                "spec": {
-                    "containers": [{
-                        "name": "gordo-deploy",
-                        "image": &format!("{}:{}", &env_config.deploy_image, &gordo.spec.deploy_version),
-                        "env": env,
-                        "resources": {
-                            "limits": {
-                                "memory": "1000Mi",
-                                "cpu": "2000m"
-                            },
-                            "requests": {
-                                "memory": "500Mi",
-                                "cpu": "250m"
-                            }
-                        }
-                    }],
-                    "restartPolicy": "Never"
-                }
-            }
-        }
-
-    });
-    let serialized_spec = serde_json::to_vec(&spec).unwrap();
+    // Job manifest for launching this gordo config into a workflow
+    let job = DeployJob::new(&gordo, &env_config);
 
     // Before launching this job, remove previous jobs for this project
     remove_gordo_deploy_jobs(&gordo, &client, &namespace);
 
     // Send off job, later we can add support to watching the job if needed via `jobs.watch(..)`
-    info!("Launching job - {}!", &job_name);
+    info!("Launching job - {}!", &job.name);
     let postparams = PostParams::default();
     let jobs = Api::v1Job(client.clone()).within(&namespace);
 
-    match jobs.create(&postparams, serialized_spec) {
+    let serialized_job_manifest = job.as_vec();
+    match jobs.create(&postparams, serialized_job_manifest) {
         Ok(job) => info!("Submitted job: {:?}", job.metadata.name),
         Err(e) => error!("Failed to submit job with error: {:?}", e),
     }
@@ -352,18 +282,4 @@ pub(crate) fn remove_gordo_deploy_jobs(gordo: &Gordo, client: &APIClient, namesp
             }),
         Err(e) => error!("Failed to list jobs: {:?}", e),
     }
-}
-
-/// Generate a name which is no greater than 63 chars in length
-/// always keeping the `prefix` and as much of `suffix` as possible, favoring its ending.
-pub fn deploy_job_name(prefix: &str, suffix: &str) -> String {
-    let suffix = suffix
-        .chars()
-        .rev()
-        .take(63 - prefix.len())
-        .collect::<Vec<char>>()
-        .iter()
-        .rev()
-        .collect::<String>();
-    format!("{}{}", prefix, suffix)
 }
